+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.13.0] - 2024-01-31
+- Add support for OpenAPI v3 deploy, update and clean-up in API Gateway
+- Implement permission setting for lambda functions in OpenAPI v3 implementations
+
 # [1.12.2] - 2024-01-31
 - Added validation for the configuration parameter `iam_suffix`
 - Added warning to logs in case of unknown parameters in the configuration file